update php7
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@37966 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/php7/main/php_scandir.c
@@ -17,8 +17,6 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id$ */
-
 #include "php.h"
 #include "php_scandir.h"
 
@@ -73,7 +71,7 @@ PHPAPI int php_scandir(const char *dirname, struct dirent **namelist[], int (*se
 	}
 
 	while (!php_readdir_r(dirp, (struct dirent *)entry, &dp) && dp) {
-		int dsize = 0;
+		size_t dsize = 0;
 		struct dirent *newdp = NULL;
 
 		if (selector && (*selector)(dp) == 0) {
@@ -95,7 +93,7 @@ PHPAPI int php_scandir(const char *dirname, struct dirent **namelist[], int (*se
 			vector = newv;
 		}
 
-		dsize = sizeof (struct dirent) + (((int)strlen(dp->d_name) + 1) * sizeof(char));
+		dsize = sizeof (struct dirent) + ((strlen(dp->d_name) + 1) * sizeof(char));
 		newdp = (struct dirent *) malloc(dsize);
 
 		if (newdp == NULL) {
